requirements.txt
Pillow>=10.4.0
PySide6>=6.8.0
Pillow>=10.4.0
PySide6>=6.8.0
